bandwidthInfo: Int too small

+++ lib/src/main/kotlin/io/ipfs/kotlin/model/BandWidthInfo.kt
@@ -1,6 +1,6 @@
 package io.ipfs.kotlin.model
 
-data class BandWidthInfo(val TotalIn: Int,
-                         val TotalOut: Int,
+data class BandWidthInfo(val TotalIn: Long,
+                         val TotalOut: Long,
                          val RateIn: Double,
                          val RateOut: Double)
